removed navigation. Less buggy now
