--- add_to_app/android_view/app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/navigation_feed"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_feed" />
-
-    <item
-        android:id="@+id/navigation_docs"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_docs" /> <!-- TODO: Find a better icon-->
-
-</menu>
