--- add_to_app/android_view/app/src/main/res/navigation/mobile_navigation.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_feed">
-
-    <fragment
-        android:id="@+id/navigation_feed"
-        android:name="dev.flutter.example.androidView.ui.feed.FeedFragment"
-        android:label="@string/title_feed"
-        tools:layout="@layout/fragment_feed" />
-
-    <fragment
-        android:id="@+id/navigation_docs"
-        android:name="dev.flutter.example.androidView.ui.docs.DocsFragment"
-        android:label="@string/title_docs"
-        tools:layout="@layout/fragment_docs" />
-
-</navigation>
